requirements.txt
tensorflow>=2.15
torch>=2.0
torchvision>=0.15
albumentations>=1.3
scikit-learn>=1.3
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
streamlit>=1.28
plotly>=5.17
streamlit-folium>=0.15
folium>=0.14
reportlab>=4.0
Pillow>=10.0
PyYAML>=6.0
pytest>=7.4
flake8>=6.1
scikit-image>=0.21
scipy>=1.10
python-dotenv>=1.0
